NegotiationEngine/API PILOT 1/requirements.txt
Flask==2.0.1
Flask_Cors==3.0.10
Flask_Login==0.5.0
geopy==2.1.0
pymongo==3.12.0
Werkzeug==2.0.1
dnspython
python-dotenv
python-dateutil
flask_expects_json
